+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Add settings to setup Joanie
 - Add course runs to the course search API
 - Add autocomplete feature on Page select in Person plugin